Revert "Fold CompletionQueue and ServerCompletionQueue into grpc_impl"
