+++ include/grpcpp/impl/codegen/async_stream.h
@@ -28,6 +28,8 @@
 
 namespace grpc {
 
+class CompletionQueue;
+
 namespace internal {
 /// Common interface for all client side asynchronous streaming.
 class ClientAsyncStreamingInterface {
