+++ include/grpcpp/impl/codegen/client_callback.h
@@ -36,6 +36,7 @@ class Channel;
 namespace grpc {
 
 class ClientContext;
+class CompletionQueue;
 
 namespace internal {
 class RpcMethod;
