+++ include/grpcpp/impl/codegen/client_unary_call.h
@@ -27,7 +27,9 @@
 
 namespace grpc {
 
+class Channel;
 class ClientContext;
+class CompletionQueue;
 
 namespace internal {
 class RpcMethod;
