+++ include/grpcpp/impl/codegen/intercepted_channel.h
@@ -21,10 +21,6 @@
 
 #include <grpcpp/impl/codegen/channel_interface.h>
 
-namespace grpc_impl {
-class CompletionQueue;
-}
-
 namespace grpc {
 
 namespace internal {
@@ -50,7 +46,7 @@ class InterceptedChannel : public ChannelInterface {
       : channel_(channel), interceptor_pos_(pos) {}
 
   Call CreateCall(const RpcMethod& method, ClientContext* context,
-                  ::grpc_impl::CompletionQueue* cq) override {
+                  CompletionQueue* cq) override {
     return channel_->CreateCallInternal(method, context, cq, interceptor_pos_);
   }
 
@@ -62,8 +58,7 @@ class InterceptedChannel : public ChannelInterface {
   }
 
   void NotifyOnStateChangeImpl(grpc_connectivity_state last_observed,
-                               gpr_timespec deadline,
-                               ::grpc_impl::CompletionQueue* cq,
+                               gpr_timespec deadline, CompletionQueue* cq,
                                void* tag) override {
     return channel_->NotifyOnStateChangeImpl(last_observed, deadline, cq, tag);
   }
@@ -72,9 +67,7 @@ class InterceptedChannel : public ChannelInterface {
     return channel_->WaitForStateChangeImpl(last_observed, deadline);
   }
 
-  ::grpc_impl::CompletionQueue* CallbackCQ() override {
-    return channel_->CallbackCQ();
-  }
+  CompletionQueue* CallbackCQ() override { return channel_->CallbackCQ(); }
 
   ChannelInterface* channel_;
   size_t interceptor_pos_;
