+++ include/grpcpp/impl/codegen/server_context.h
@@ -43,12 +43,12 @@ struct census_context;
 
 namespace grpc_impl {
 
-class CompletionQueue;
 class Server;
 }  // namespace grpc_impl
 namespace grpc {
 class ClientContext;
 class GenericServerContext;
+class CompletionQueue;
 class ServerInterface;
 template <class W, class R>
 class ServerAsyncReader;
@@ -90,7 +90,6 @@ class Call;
 class ServerReactor;
 }  // namespace internal
 
-class ServerInterface;
 namespace testing {
 class InteropServerContextInspector;
 class ServerContextTestSpouse;
@@ -355,7 +354,7 @@ class ServerContext {
 
   gpr_timespec deadline_;
   grpc_call* call_;
-  ::grpc_impl::CompletionQueue* cq_;
+  CompletionQueue* cq_;
   bool sent_initial_metadata_;
   mutable std::shared_ptr<const AuthContext> auth_context_;
   mutable internal::MetadataMap client_metadata_;
