+++ include/grpcpp/impl/codegen/service_type.h
@@ -29,11 +29,12 @@
 namespace grpc_impl {
 
 class Server;
-class CompletionQueue;
 }  // namespace grpc_impl
 namespace grpc {
 
+class CompletionQueue;
 class ServerInterface;
+class ServerCompletionQueue;
 class ServerContext;
 
 namespace internal {
