+++ include/grpcpp/server_builder.h
@@ -21,6 +21,13 @@
 
 #include <grpcpp/server_builder_impl.h>
 
+namespace grpc_impl {
+
+class Server;
+class ServerCredentials;
+class ResourceQuota;
+}  // namespace grpc_impl
+
 namespace grpc {
 
 typedef ::grpc_impl::ServerBuilder ServerBuilder;
