+++ test/cpp/codegen/compiler_test_golden
@@ -42,8 +42,6 @@
 
 namespace grpc_impl {
 class Channel;
-class CompletionQueue;
-class ServerCompletionQueue;
 }  // namespace grpc_impl
 
 namespace grpc {
@@ -51,9 +49,8 @@ namespace experimental {
 template <typename RequestT, typename ResponseT>
 class MessageAllocator;
 }  // namespace experimental
-}  // namespace grpc_impl
-
-namespace grpc {
+class CompletionQueue;
+class ServerCompletionQueue;
 class ServerContext;
 }  // namespace grpc
